% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation.R
\name{bjerksund_put}
\alias{bjerksund_put}
\title{Bjerksund–Stensland 2002 American put (via put‑call symmetry)}
\usage{
bjerksund_put(S, K, r, q, T, sigma)
}
\description{
Bjerksund–Stensland 2002 American put (via put‑call symmetry)
}
