% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{date_to_ttm}
\alias{date_to_ttm}
\title{Convert dates to time‑to‑maturity (always calendar years)}
\usage{
date_to_ttm(expiry, valuation_date = Sys.Date())
}
\arguments{
\item{expiry}{Date, character (YYYY-MM-DD) or numeric (if numeric, returned unchanged)}

\item{valuation_date}{Date, character (YYYY-MM-DD) or missing (uses Sys.Date)}
}
\value{
Time in years (calendar)
}
\description{
Given an expiry, passed as a numeric (fraction of a year of 365 days or as a Date object), and an optional valuation date as a
Date object (by default the current system date will be used) the function returns the time to maturity of the option as a fraction of
the calendar year (365 days)
}
\examples{
date_to_ttm(0.5)
#option expiry
option_maturity <- as.Date("2026-04-17")
#valuation day
today <- "2026-03-08"
date_to_ttm(option_maturity, today)

}
