% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_global_r}
\alias{get_global_r}
\title{Get global risk‑free rate}
\usage{
get_global_r()
}
\value{
Numeric, the current global risk-free rate (e.g. 0.05 for 5\%)
}
\description{
Get global risk‑free rate
}
\examples{
# Get current global rate
get_global_r()

# Set a new rate and verify
old_rate <- get_global_r()
set_global_r(0.03)
get_global_r()
set_global_r(old_rate) # restore

}
\seealso{
\code{\link{set_global_r}} to set risk-free rate.
}
