% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_min_reliable_ttm}
\alias{get_min_reliable_ttm}
\title{Get minimum reliable TTM}
\usage{
get_min_reliable_ttm()
}
\value{
Numeric The minimum time to maturity value currently in use
}
\description{
OptionSuite relies, for several option valuation calculations, on publicly available libraries like QuantLib. During development, empirically,
it has been discovered that these algorithms do not provide valid results for time to maturity values below a certain threshold value (around)
0.0014. It is then advisable to globally set the minimal time to maturity level used by OptionsSuite calculation to this small value not to
run into the risk of obtaining unreliable values. But the user is free to bypass this constraint, under her own responsibility.
}
\details{
This function shows the minimal time to maturity currently in use.
}
\examples{
old_min_ttm <- get_min_reliable_ttm()
set_min_reliable_ttm(0.025)
get_min_reliable_ttm()
set_min_reliable_ttm(old_min_ttm) # restore

}
\seealso{
\code{\link{set_min_reliable_ttm}} to modify this value.
}
