% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_use_commission}
\alias{get_use_commission}
\title{Get current commission setting}
\usage{
get_use_commission()
}
\value{
Logical, TRUE if commissions are enabled
}
\description{
Get current commission setting
}
\examples{
get_use_commission()
}
