% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legs.R
\name{leg.get_direction}
\alias{leg.get_direction}
\title{Get direction from a leg}
\usage{
leg.get_direction(leg)
}
\arguments{
\item{leg}{An object of class "leg"}
}
\value{
Character "buy" or "sell"
}
\description{
Get direction from a leg
}
\examples{
opt <- make_option(right = "c", strike = 100, expiry = "2026-12-31",
                  entry_price = 10, underlying_at_entry = 100)
leg <- make_leg(opt, quantity = 10, direction = "buy")
leg.get_direction(leg)

}
