% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legs.R
\name{leg.get_instrument}
\alias{leg.get_instrument}
\title{Get instrument from a leg}
\usage{
leg.get_instrument(leg)
}
\arguments{
\item{leg}{An object of class "leg"}
}
\value{
The underlying instrument object
}
\description{
Get instrument from a leg
}
\examples{
opt <- make_option(right = "c", strike = 100, expiry = "2026-12-31",
                  entry_price = 10, underlying_at_entry = 100)
leg <- make_leg(opt, quantity = 10, direction = "buy")
leg.get_instrument(leg)

}
