% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legs.R
\name{leg.get_multiplier}
\alias{leg.get_multiplier}
\title{Get effective multiplier for a leg}
\usage{
leg.get_multiplier(leg)
}
\arguments{
\item{leg}{An object of class "leg"}
}
\value{
Numeric multiplier
}
\description{
The effective multiplier is the instrument's multiplier (100 for options, 1 for underlyings)
}
\examples{
opt <- make_option(right = "c", strike = 100, expiry = "2026-12-31",
                  multiplier = 100, entry_price = 10, underlying_at_entry = 100)
leg <- make_leg(opt, quantity = 10, direction = "buy")
leg.get_multiplier(leg)  # Returns 100

}
