% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{option-getters}
\alias{option-getters}
\alias{option.get_right}
\alias{option.get_strike}
\alias{option.get_expiry}
\alias{option.get_style}
\alias{option.get_multiplier}
\alias{option.get_commission}
\alias{option.get_entry_price}
\alias{option.get_iv}
\title{Get option attributes}
\usage{
option.get_right(object)

option.get_strike(object)

option.get_expiry(object)

option.get_style(object)

option.get_multiplier(object)

option.get_commission(object)

option.get_entry_price(object)

option.get_iv(object)
}
\arguments{
\item{object}{An object of class "option"}
}
\value{
The requested attribute value
}
\description{
These functions extract specific attributes from an option object.
They are simple accessors that return the stored value.
}
