% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{option.extrinsic}
\alias{option.extrinsic}
\title{Extrinsic value = total value - intrinsic}
\usage{
option.extrinsic(
  object,
  underlying,
  ttm,
  volatility = NULL,
  r = NULL,
  q = 0,
  ...
)
}
\description{
Extrinsic value = total value - intrinsic
}
