% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{option.value}
\alias{option.value}
\title{Value of an option}
\usage{
option.value(object, underlying, ttm, volatility = NULL, r = NULL, q = 0, ...)
}
\arguments{
\item{object}{An object of class "option"}

\item{underlying}{Spot price of underlying}

\item{ttm}{Time to maturity in years}

\item{volatility}{Volatility (if NULL, uses stored iv)}

\item{r}{Risk‑free rate (if NULL, uses global)}

\item{q}{Dividend yield (default 0)}

\item{...}{Additional arguments passed to price_option}
}
\value{
Numeric value
}
\description{
Value of an option
}
