% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation.R
\name{price_option}
\alias{price_option}
\title{Price an option (European or American) and return full Greeks}
\usage{
price_option(
  style = c("e", "a"),
  type = c("c", "p"),
  S,
  K,
  r = NULL,
  q = 0,
  T,
  iv,
  ...
)
}
\arguments{
\item{style}{"e" for European" or "a" for American}

\item{type}{"c" for call or "p" for put}

\item{S}{Spot}

\item{K}{Strike}

\item{r}{Rate (NULL uses global)}

\item{q}{Dividend yield (0 by default)}

\item{T}{Time to maturity}

\item{iv}{Implied volatility}

\item{...}{Additional arguments passed to specific methods}
}
\value{
List with value and Greeks
}
\description{
Valuation is provided by applying the Black-Scholes model for European Options through the QuantLib public library
or the Bjerksund–Stensland (2002) closed model for American options through the fOptions implementation.
Given the usual input parameters listed below, the returned object is a list containing option value at the given
time to maturity and volatility levels, plus all the first-order greeks: delta, vega, theta, gamma, and rho.
}
\examples{
# As of March 8th 2026, BMY Apr17Call60 is quoted on TWS 2.36x2.51, while BMY quotes 60.22
# ATM Implied volatility is reported as 28.8\%. The very same data are confirmed on ToS, besides
# BMY being quoted 60.29 and the bid-ask values for the option being 2.31x2.51
# Dividend yield is indicated as 4.2\% and interest rate as 4.5\%

get_day_count_convention()
expiry <- as.Date("2026-04-17")
valuation_date <- as.Date("2026-03-08")
ttm <- date_to_ttm(expiry, valuation_date)
price_option(style = "a",
             type = "c",
             S = 60.22,
             K = 60,
             r = 0.045,
             q = 0.042,
             T = ttm,
             iv = 0.288)
             
# As of March 8th 2026, SPX May14Put6740 is quoted on TWS 228.10x234.10, while S&P 500 quotes 6738.15
# ATM implied volatility is reported as 20.5\% while VIX is 29.49. ToS reports IV of 22.40\%.
# These options are European style with no dividends

expiry <- as.Date("2026-05-14")
valuation_date <- as.Date("2026-03-08")
ttm <- date_to_ttm(expiry, valuation_date)
price_option(style = "e",
             type = "p",
             S = 6738.15,
             K = 6740,
             r = 0.045,
             q = 0,
             T = ttm,
             iv = 0.224)

}
