% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_day_count_convention}
\alias{set_day_count_convention}
\title{Set day‑count convention for volatility annualization.}
\usage{
set_day_count_convention(convention = c("calendar/365", "trading/252"))
}
\arguments{
\item{convention}{Character, either "calendar/365" or "trading/252"}
}
\description{
Different brokers use different conventions. For instance Think Or Swim (Schwab) apparently uses the 252 days convention
(trading sessions in a year, on average) while IBKR (TWS) apparently uses 365 days (approximate days in one year) for
implied volatility and option pricing calculations. OptionSuite allows the user to adopt one or the other scheme, interchangeably.
}
\examples{
set_day_count_convention("calendar/365")
get_day_count_convention()
set_day_count_convention("trading/252")
get_day_count_convention()

}
\seealso{
\code{\link{get_day_count_convention}} to see how this parameter is currently set.
}
