% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_global_r}
\alias{set_global_r}
\title{Set global risk‑free rate}
\usage{
set_global_r(rate)
}
\arguments{
\item{rate}{Numeric, annualised rate (e.g., 0.05 for 5\%)}
}
\description{
Set global risk‑free rate
}
\examples{
get_global_r()
set_global_r(0.044)
get_global_r()

}
\seealso{
\code{\link{get_global_r}} to see the currently used risk-free rate.
}
