% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_min_reliable_ttm}
\alias{set_min_reliable_ttm}
\title{Set minimum reliable TTM for numerical pricing engines}
\usage{
set_min_reliable_ttm(ttm)
}
\arguments{
\item{ttm}{Numeric, positive the choosen minimal time to maturity value use for option value calculations}
}
\description{
OptionSuite relies, for several option valuation calculations, on publicly available libraries like QuantLib. During development, empirically,
it has been discovered that these algorithms do not provide valid results for time to maturity values below a certain threshold value (around)
0.0014. It is then advisable to globally set the minimal time to maturity level used by OptionsSuite calculation to this small value not to
run into the risk of obtaining unreliable values. But the user is free to bypass this constraint, under her own responsibility.
}
\examples{
old_min_ttm <- get_min_reliable_ttm
set_min_reliable_ttm(0.025)
get_min_reliable_ttm()
set_min_reliable_ttm(old_min_ttm) # restore

}
\seealso{
\code{\link{get_min_reliable_ttm}} to see how this value is currently set.
}
