% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_use_commission}
\alias{set_use_commission}
\title{Set whether to include commission in calculations. At startup the Suite ignores
commissions so a call to this function is needed to change the default behavior
of the suite.}
\usage{
set_use_commission(use = TRUE)
}
\arguments{
\item{use}{Logical, TRUE to include commissions (DEFAULT), FALSE to ignore.}
}
\description{
Set whether to include commission in calculations. At startup the Suite ignores
commissions so a call to this function is needed to change the default behavior
of the suite.
}
\examples{
set_use_commission(TRUE)

}
