% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{strategy.get_legs}
\alias{strategy.get_legs}
\title{Get legs from a strategy}
\usage{
strategy.get_legs(strategy)
}
\arguments{
\item{strategy}{An object of class "strategy"}
}
\value{
List of leg objects
}
\description{
Get legs from a strategy
}
\examples{
covered_call <- make_strategy(list(leg1, leg2))
legs <- strategy.get_legs(covered_call)

}
