% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{strategy.get_name}
\alias{strategy.get_name}
\title{Get strategy name}
\usage{
strategy.get_name(strategy)
}
\arguments{
\item{strategy}{An object of class "strategy"}
}
\value{
Character strategy name
}
\description{
Get strategy name
}
\examples{
covered_call <- make_strategy(list(leg1, leg2), "Covered Call")
strategy.get_name(covered_call)

}
