% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{strategy.n_legs}
\alias{strategy.n_legs}
\title{Get number of legs in a strategy}
\usage{
strategy.n_legs(strategy)
}
\arguments{
\item{strategy}{An object of class "strategy"}
}
\value{
Integer number of legs
}
\description{
Get number of legs in a strategy
}
\examples{
covered_call <- make_strategy(list(leg1, leg2))
strategy.n_legs(covered_call)  # Returns 2

}
