% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{strategy.set_name}
\alias{strategy.set_name}
\title{Set strategy name}
\usage{
strategy.set_name(strategy, name)
}
\arguments{
\item{strategy}{An object of class "strategy"}

\item{name}{New name for the strategy}
}
\value{
Modified strategy object (invisibly)
}
\description{
Set strategy name
}
\examples{
covered_call <- make_strategy(list(leg1, leg2))
strategy.set_name(covered_call, "My Covered Call")

}
