% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{underlying.analytics}
\alias{underlying.analytics}
\title{Complete analytics for underlying}
\usage{
underlying.analytics(object, price, ...)
}
\description{
Complete analytics for underlying
}
\keyword{internal}
