% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{underlying.delta}
\alias{underlying.delta}
\title{Delta of underlying (always 1 per share)}
\usage{
underlying.delta(object, ...)
}
\description{
Delta of underlying (always 1 per share)
}
\keyword{internal}
