% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{underlying.entry_cost}
\alias{underlying.entry_cost}
\title{Entry cost for underlying position (per share)}
\usage{
underlying.entry_cost(object)
}
\description{
Entry cost for underlying position (per share)
}
\keyword{internal}
