% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instruments.R
\name{underlying.value}
\alias{underlying.value}
\title{Value of an underlying instrument (per share)}
\usage{
underlying.value(object, price)
}
\arguments{
\item{object}{An object of class "underlying"}

\item{price}{Current market price}
}
\value{
Numeric value (per share)
}
\description{
Value of an underlying instrument (per share)
}
\keyword{internal}
