% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{date_to_ttm}
\alias{date_to_ttm}
\title{Convert dates to time‑to‑maturity (always calendar years)}
\usage{
date_to_ttm(expiry, valuation_date = Sys.Date())
}
\arguments{
\item{expiry}{Date or numeric (if numeric, returned unchanged)}

\item{valuation_date}{Date (ignored if expiry is numeric)}
}
\value{
Time in years (calendar)
}
\description{
Convert dates to time‑to‑maturity (always calendar years)
}
