% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuation.R
\name{price_american}
\alias{price_american}
\title{Price an American option and return full Greeks}
\usage{
price_american(
  type = c("c", "p"),
  S,
  K,
  r = NULL,
  q,
  T,
  sigma,
  use_min_ttm = FALSE
)
}
\arguments{
\item{type}{"c" for call or "p" for put}

\item{S}{Spot}

\item{K}{Strike}

\item{r}{Risk‑free rate (if NULL, uses global)}

\item{q}{Dividend yield}

\item{T}{Time to maturity}

\item{sigma}{Volatility}

\item{use_min_ttm}{Logical; if TRUE and T < min_reliable, use min_reliable for Greeks}
}
\value{
List with value, delta, gamma, vega, theta, rho
}
\description{
Price an American option and return full Greeks
}
